Move + simplify *Requests filters
1. Deprecated filters were put into their own file
2. Filter methods were made... methods, instead of functions - this
   helps visually determine if a filter method is used and by whom.
3. Unused deprecated filter methods were removed.
