+++ amy/workshops/filters.py
@@ -3,7 +3,6 @@ import re
 
 from dal import autocomplete
 from dal_select2.widgets import (
-    ListSelect2,
     Select2,
     Select2Multiple,
     ModelSelect2Multiple,
@@ -21,17 +20,10 @@ from workshops.models import (
     Person,
     Badge,
     Airport,
-    EventRequest,
     Tag,
     Task,
     Award,
-    InvoiceRequest,
-    EventSubmission,
-    DCSelfOrganizedEventRequest,
-    TrainingRequest,
     Membership,
-    Curriculum,
-    WorkshopRequest,
 )
 
 
@@ -227,6 +219,7 @@ def filter_active_memberships_only(queryset, name, active):
                                agreement_end__gte=today)
     else:
         return queryset
+
 
 def filter_training_seats_only(queryset, name, seats):
     """Limit Memberships to only entries with some training seats allowed."""
@@ -508,125 +501,6 @@ class TraineeFilter(AMYFilterSet):
         ]
 
 
-def filter_matched(queryset, name, choice):
-    if choice == '':
-        return queryset
-    elif choice == 'u':  # unmatched
-        return queryset.filter(person=None)
-    elif choice == 'p':  # matched trainee, unmatched training
-        return queryset.filter(person__isnull=False)\
-                       .exclude(person__task__role__name='learner',
-                                person__task__event__tags__name='TTT')\
-                       .distinct()
-    else:  # choice == 't' <==> matched trainee and training
-        return queryset.filter(person__task__role__name='learner',
-                               person__task__event__tags__name='TTT')\
-                       .distinct()
-
-
-def filter_by_person(queryset, name, value):
-    if value == '':
-        return queryset
-    else:
-        # 'Harry Potter' -> ['Harry', 'Potter']
-        tokens = re.split('\s+', value)
-        # Each token must match email address or github username or personal or
-        # family name.
-        for token in tokens:
-            queryset = queryset.filter(
-                Q(personal__icontains=token) |
-                Q(middle__icontains=token) |
-                Q(family__icontains=token) |
-                Q(email__icontains=token) |
-                Q(person__personal__icontains=token) |
-                Q(person__middle__icontains=token) |
-                Q(person__family__icontains=token) |
-                Q(person__email__icontains=token)
-            )
-        return queryset
-
-
-def filter_affiliation(queryset, name, affiliation):
-    if affiliation == '':
-        return queryset
-    else:
-        return queryset.filter(Q(affiliation__icontains=affiliation) |
-                               Q(person__affiliation__icontains=affiliation)) \
-                       .distinct()
-
-
-def filter_training_requests_by_state(queryset, name, choice):
-    if choice == 'no_d':
-        return queryset.exclude(state='d')
-    else:
-        return queryset.filter(state=choice)
-
-
-def filter_non_null_manual_score(queryset, name, manual_score):
-    if manual_score:
-        return queryset.filter(score_manual__isnull=False)
-    return queryset
-
-
-class TrainingRequestFilter(AMYFilterSet):
-    search = django_filters.CharFilter(
-        label='Name or Email',
-        method=filter_by_person,
-    )
-
-    group_name = django_filters.CharFilter(
-        field_name='group_name',
-        lookup_expr='icontains',
-        label='Group')
-
-    state = django_filters.ChoiceFilter(
-        label='State',
-        choices=(('no_d', 'Pending or accepted'),) + TrainingRequest.STATE_CHOICES,
-        method=filter_training_requests_by_state,
-    )
-
-    matched = django_filters.ChoiceFilter(
-        label='Is Matched?',
-        choices=(
-            ('', 'Unknown'),
-            ('u', 'Unmatched'),
-            ('p', 'Matched trainee, unmatched training'),
-            ('t', 'Matched trainee and training'),
-        ),
-        method=filter_matched,
-    )
-
-    nonnull_manual_score = django_filters.BooleanFilter(
-        label='Manual score applied',
-        method=filter_non_null_manual_score,
-        widget=widgets.CheckboxInput,
-    )
-
-    affiliation = django_filters.CharFilter(
-        method=filter_affiliation,
-    )
-
-    location = django_filters.CharFilter(lookup_expr='icontains')
-
-    order_by = NamesOrderingFilter(
-        fields=(
-            'created_at',
-            'score_total',
-        ),
-    )
-
-    class Meta:
-        model = TrainingRequest
-        fields = [
-            'search',
-            'group_name',
-            'state',
-            'matched',
-            'affiliation',
-            'location',
-        ]
-
-
 class TaskFilter(AMYFilterSet):
     event = django_filters.ModelChoiceFilter(
         queryset=Event.objects.all(),
@@ -712,137 +586,3 @@ class BadgeAwardsFilter(AMYFilterSet):
         fields = (
             'awarded_after', 'awarded_before', 'event',
         )
-
-
-class EventRequestFilter(AMYFilterSet, StateFilterSet):
-    assigned_to = ForeignKeyAllValuesFilter(Person, widget=Select2())
-    country = AllCountriesFilter(widget=Select2())
-    workshop_type = django_filters.ChoiceFilter(
-        choices=(('swc', 'Software-Carpentry'),
-                 ('dc', 'Data-Carpentry')),
-        label='Workshop type',
-        empty_label='All',
-    )
-
-    order_by = django_filters.OrderingFilter(
-        fields=(
-            'created_at',
-        ),
-    )
-
-    class Meta:
-        model = EventRequest
-        fields = [
-            'state',
-            'assigned_to',
-            'workshop_type',
-            'country',
-        ]
-
-
-class WorkshopRequestFilter(AMYFilterSet, StateFilterSet):
-    assigned_to = ForeignKeyAllValuesFilter(Person, widget=Select2())
-    country = AllCountriesFilter(widget=Select2())
-    requested_workshop_types = django_filters.ModelMultipleChoiceFilter(
-        label='Requested workshop types',
-        queryset=Curriculum.objects.all(),
-        widget=widgets.CheckboxSelectMultiple(),
-    )
-
-    order_by = django_filters.OrderingFilter(
-        fields=(
-            'created_at',
-        ),
-    )
-
-    class Meta:
-        model = WorkshopRequest
-        fields = [
-            'state',
-            'assigned_to',
-            'requested_workshop_types',
-            'country',
-        ]
-
-
-class InvoiceRequestFilter(AMYFilterSet):
-    STATUS_CHOICES = (('', 'All'),) + InvoiceRequest.STATUS_CHOICES
-    status = django_filters.ChoiceFilter(
-        choices=STATUS_CHOICES,
-    )
-
-    organization = django_filters.ModelChoiceFilter(
-        queryset=Organization.objects.all(),
-        label='Organization',
-        widget=autocomplete.ModelSelect2(
-            url='organization-lookup',
-            attrs=SIDEBAR_DAL_WIDTH,
-        ),
-    )
-
-    order_by = django_filters.OrderingFilter(
-        fields=(
-            'event__slug',
-            'organization__domain',
-        ),
-    )
-
-    class Meta:
-        model = InvoiceRequest
-        fields = [
-            'status',
-            'organization',
-        ]
-
-
-def filter_active_eventsubmission(qs, name, value):
-    if value == 'true':
-        return qs.filter(active=True)
-    elif value == 'false':
-        return qs.filter(active=False)
-    return qs
-
-
-class EventSubmissionFilter(AMYFilterSet, StateFilterSet):
-    assigned_to = ForeignKeyAllValuesFilter(Person, widget=Select2())
-
-    order_by = django_filters.OrderingFilter(
-        fields=(
-            'created_at',
-        ),
-    )
-
-    class Meta:
-        model = EventSubmission
-        fields = [
-            'state',
-            'assigned_to',
-        ]
-
-
-def filter_active_dcselforganizedeventrequest(qs, name, value):
-    if value == 'true':
-        return qs.filter(active=True)
-    elif value == 'false':
-        return qs.filter(active=False)
-    return qs
-
-
-class DCSelfOrganizedEventRequestFilter(AMYFilterSet, StateFilterSet):
-    assigned_to = ForeignKeyAllValuesFilter(Person, widget=Select2())
-
-    order_by = django_filters.OrderingFilter(
-        fields=(
-            'created_at',
-        ),
-    )
-
-    class Meta:
-        model = DCSelfOrganizedEventRequest
-        fields = [
-            'state',
-            'assigned_to',
-        ]
-        order_by = [
-            '-created_at', 'created_at',
-        ]
